requirements.txt
ailment==8.18.10.25
amqp==2.3.2
ana==0.5
angr==8.18.10.25
archinfo==8.18.10.25
billiard==3.5.0.5
bitstring==3.1.5
cachetools==3.0.0
capstone==3.0.5.post1
celery==4.2.1
cffi==1.11.5
claripy==8.18.10.25
cle==8.18.10.25
colorama==0.4.1
cooldict==1.4
cycler==0.10.0
decorator==4.3.0
dpkt==1.9.1
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
idalink==0.12
itanium-demangler==1.0
kiwisolver==1.0.1
kombu==4.2.1
matplotlib==3.0.2
mulpyplexer==0.8
networkx==2.2
numpy==1.15.4
pefile==2018.8.8
plumbum==1.6.7
progressbar==2.5
pycparser==2.19
pyelftools==0.25
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pyvex==8.18.10.25
redis==3.0.1
rpyc==4.0.2
six==1.11.0
smmap2==2.0.5
sortedcontainers==2.1.0
tqdm==4.28.1
unicorn==1.0.1
vine==1.1.4
z3-solver==4.5.1.0.post2
